Fix package import due to move
