+++ plotly/tests/test_plot_ly/test_api/test_v1/test_utils.py
@@ -10,7 +10,7 @@ from plotly.api.utils import to_native_utf8_string
 from plotly.api.v1 import utils
 from plotly.exceptions import PlotlyError, PlotlyRequestError
 from plotly.session import sign_in
-from plotly.tests.test_core.test_api import PlotlyApiTestCase
+from plotly.tests.test_plot_ly.test_api import PlotlyApiTestCase
 from plotly.tests.utils import PlotlyTestCase
 
 import sys
