+++ plotly/tests/test_plot_ly/test_api/test_v2/test_folders.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 
 from plotly.api.v2 import folders
-from plotly.tests.test_core.test_api import PlotlyApiTestCase
+from plotly.tests.test_plot_ly.test_api import PlotlyApiTestCase
 
 
 class FoldersTest(PlotlyApiTestCase):
