+++ plotly/tests/test_plot_ly/test_api/test_v2/test_users.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 
 from plotly.api.v2 import users
-from plotly.tests.test_core.test_api import PlotlyApiTestCase
+from plotly.tests.test_plot_ly.test_api import PlotlyApiTestCase
 
 
 class UsersTest(PlotlyApiTestCase):
